Rename 1-s2.0-S2213138821005750-main.pdf to Artificial intelligence assisted technoeconomic optimization scenarios of hybrid energy systems for water management of an isolated community
Water is an essential resource demanded worldwide and it is quite debatable owing to the economic, political,
and energy characteristics of any region. Off-grid water filtration plants are an alternative for communities where
transportation of freshwater becomes a real challenge due to a lack of infrastructure for the water potabilization
processes. For such potable water filtration plants, hybrid renewable energy systems (HRES) can be a viable
solution to meet their energy demand meanwhile providing a sustainable water solution.
The main contribution of this work is the unique methodology, which starts with a sizing procedure of various
hybrid energy systems using a commercial software “Hybrid Optimization of Multiple Energy Resources (HOMER)”
and spreadsheet algorithms, followed by a “Non-dominating Sorting Genetic Algorithm II (NSGA-II)” based multiobjective
optimization. Single-objective optimization scenarios contain photovoltaic installation capacity, wind
turbines, diesel generators, and battery energy storage systems including Pb-acid (Lead-acid), Li-ion (Lithium-ion),
and AGM (Absorbent Glass Mat) technologies as design variables to maximize the cost of electricity or net-present-cost.
Multiobjective optimization also involved environmental (CO2 emissions i.e. carbon dioxide emissions) and water cost
indices as an additional packet to single-objective optimization scenarios. Afterward, a multicriteria decision-making
tool using “The Order of Preference by Similarity to Ideal Solution (TOPSIS)” is applied on the Pareto front to attain the
final optimization results. The analysis is further explored in depth by generating digital twins (surrogate or meta
model) of HRES data using artificial intelligence techniques (artificial neural network and group-method-of-datahandling).
Furthermore, calculus and statistical sensitivity analysis assist in the identification of the significant variables
in the design procedure. In summary, the technical contribution of this work can be divided into two sections.
The first one is the design of a hybrid energy system for the water management of an isolated community of the
indigenous Mayan region of Yucatan, Mexico, which has never been considered before. Secondly, the technical
contribution is related to the usage of environmental emissions as an objective function, which is not considered in the
traditional design of hybrid energy systems by the software HOMER. Environmental emission as an objective function
is not considered while designing a hybrid energy system in commerical softwares like HOMER, in fact, HOMER
provides a list of environmental impacts but it is a secondary outcome as a result of technoeconomic optimization.
Analysis of results between HOMER pro and spreadsheet has shown conformity, reporting that the optimal case
consists of a photovoltaic system, diesel generator, and Li-ion technology of battery storage with capacities of ~17 kW,
~5kW, and 44–48 kWh, respectively, corresponding to a net present cost ranging from 70,000 United States Dollars
(USD) to 79,000 USD and a cost of electricity ranging from 0.205 to 0.229 USD/kWh. The achievements obtained with
multiobjective optimization indicate that the cost of electricity and net present cost can be further reduced by 0.86 %
and 0.73 %, respectively, at a decrement of only 0.4% of the renewable fraction as compared to the single objective
optimization scenario. It is concluded that multiobjective optimization provides an add-in feature to HOMER by using
environmental emissions as an objective function.
The design procedure and adapted methodology can be useful to promote sustainable development in the
statewide context and can provide a scientific justification to national energy policymakers.