Global sensitivity analysis of a generator-absorber heat exchange (GAX) system’s thermal performance with a hybrid energy source: An approach using artificial intelligence models
Generator-absorber heat exchange (GAX) systems represent a promising alternative to substitute environmentally harmful refrigeration devices based on conventional vapor compression, as long as a proper analysis of thermal performance and the complex interactions of heat transfer that occur into GAX cycle is taken in consideration. In this research, a cooling process based on a GAX system that uses ammonia-water working fluid and a hybrid source (natural gas-solar) is studied to analyze the variables that affect the system’s thermal performance. The work’s novelty is the hybridization between artificial intelligence (AI) modeling and the global sensitivity analysis (GSA) developed with the PAWN method. Experimental data was obtained from a system with a cooling capacity of 10.5 kW (3 Ton), designed to work at heat source temperatures of 200 ◦C. The measured variables were the temperatures at generator, heat at evaporator, and working fluid volumetric flow. Three AI techniques (artificial neural networks, genetic programming, and support vector machines) were evaluated for modeling the thermodynamic cycle. Results obtained from the PAWN method applied to the artificial neural network, since it was the best AI model, indicates that the operational parameters with a greater impact in the system’s performance are the inlet temperature at the generator (30.7 %) and the heat measured at the evaporator for NH3 (27.4 %), for the first output COPNH3. For the second output COPH2O, the inlet temperature
at the generator (32.5 %) and the and heat measured at the evaporator for H2O (26.7 %), have a greater impact for such output. The proposed IA-GSA methodology contributes to the development of operational decision-making related to instrumentation, operation performance, and corrective and/or preventive maintenance
actions of GAX systems. The developed thermal performance model has potential for implementation in embedded systems (smart sensors) as a critical element in control and optimization strategies to improve the performance of these cycles.